fix some stuff related to changing num_frame_params
The lower bound for skip frame (apparently) wasn't low enough ---
inconsistent feasibility being returned even for exactly the same
parameters.  Decreased it as per this change, and it seemed to fix it.

The second change is related to frame params.   Frame params are floats
that are truncated to get the actual parameters for the script.  When
initializing them as part of a feasibility run a random normal
perturbation centred around 0 is added to the base values.  This change
adds 0.5 to this random value so that small values do not effect the
frame times ultimately used in the script.